Add test for Task6, fix type in interface

+++ tyuiu.cources.programming.interfaces/ISprint3Task6.cs
@@ -8,37 +8,37 @@ namespace tyuiu.cources.programming.interfaces.Sprint3
 {
     internal interface ISprint3Task6
     {
-        public interface ISprint3Task6V0 { double  GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V1 { double  GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V2 { double  GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V3 { double  GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V4 { double  GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V5 { double  GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V6 { double  GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V7 { double  GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V8 { double  GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V9 { double  GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V10 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V11 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V12 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V13 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V14 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V15 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V16 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V17 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V18 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V19 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V20 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V21 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V22 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V23 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V24 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V25 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V26 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V27 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V28 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V29 { double GetDivider(int startValue, int stopValue); }
-        public interface ISprint3Task6V30 { double GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V0 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V1 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V2 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V3 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V4 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V5 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V6 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V7 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V8 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V9 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V10 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V11 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V12 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V13 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V14 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V15 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V16 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V17 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V18 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V19 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V20 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V21 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V22 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V23 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V24 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V25 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V26 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V27 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V28 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V29 { int GetDivider(int startValue, int stopValue); }
+        public interface ISprint3Task6V30 { int GetDivider(int startValue, int stopValue); }
 
     }
 }
